requirements.txt
fastapi
uvicorn
cryptography
argon2-cffi
python-multipart
motor
pydantic
